Use ExecReload and ExecStartPost instead of path unit for more reliability

+++ install.sh
@@ -1,34 +1,29 @@
-#!/bin/sh
+#!/usr/bin/env sh
 
 if [ "$(whoami)" != "root" ]; then
   echo Script must be run as root
   exit 1
 fi
 
-if [ -e "/usr/local/bin/uninstall-pve-cert-copy.sh" ]; then
-  /usr/local/bin/uninstall-pve-cert-copy.sh
-fi
+# Uninstall old version
+[ -e /usr/local/bin/uninstall-pve-cert-copy.sh ] && /usr/local/bin/uninstall-pve-cert-copy.sh
 
-cat << EOF > /etc/systemd/system/pve-cert-copy.service
-[Unit]
-Description=Copy the certificate from Proxmox Virtual Environment to Proxmox Backup Server
-[Service]
-Type=oneshot
-ExecStart=/usr/bin/cp /etc/pve/nodes/$(hostname)/pveproxy-ssl.pem /etc/proxmox-backup/proxy.pem
-ExecStart=/usr/bin/cp /etc/pve/nodes/$(hostname)/pveproxy-ssl.key /etc/proxmox-backup/proxy.key
-ExecStart=/usr/bin/chmod 640 /etc/proxmox-backup/proxy.key /etc/proxmox-backup/proxy.pem
-ExecStart=/usr/bin/chgrp backup /etc/proxmox-backup/proxy.key /etc/proxmox-backup/proxy.pem
-ExecStart=/usr/bin/systemctl reload proxmox-backup-proxy.service
+mkdir --parents /etc/systemd/system/pveproxy.service.d
+
+cat << EOF > /usr/local/bin/pve-cert-copy.sh
+#!/usr/bin/env sh
+[ ! -e /etc/proxmox-backup ] && exit 0
+cp --preserve=timestamps /etc/pve/local/pveproxy-ssl.key /etc/proxmox-backup/proxy.key
+cp --preserve=timestamps /etc/pve/local/pveproxy-ssl.pem /etc/proxmox-backup/proxy.pem
+chmod 640 /etc/proxmox-backup/proxy.key /etc/proxmox-backup/proxy.pem
+chgrp backup /etc/proxmox-backup/proxy.key /etc/proxmox-backup/proxy.pem
+systemctl reload proxmox-backup-proxy.service
 EOF
 
-cat << EOF > /etc/systemd/system/pve-cert-copy.path
-[Unit]
-Description=Watch the certificate from Proxmox Virtual Environment to Proxmox Backup Server
-[Path]
-PathChanged=/etc/pve/nodes/$(hostname)/pveproxy-ssl.pem
-Unit=pve-cert-copy.service
-[Install]
-WantedBy=multi-user.target
+cat << EOF > /etc/systemd/system/pveproxy.service.d/pve-cert-copy-override.conf
+[Service]
+ExecReload=-/usr/local/bin/pve-cert-copy.sh
+ExecStartPost=-/usr/local/bin/pve-cert-copy.sh
 EOF
 
 cat << EOF > /usr/local/bin/uninstall-pve-cert-copy.sh
@@ -37,16 +32,16 @@ if [ "\$(whoami)" != "root" ]; then
   echo Script must be run as root
   exit 1
 fi
-systemctl disable --now pve-cert-copy.path
+rm /etc/systemd/system/pveproxy.service.d/pve-cert-copy-override.conf
+[ \$(ls -A /etc/systemd/system/pveproxy.service.d) ] || rmdir /etc/systemd/system/pveproxy.service.d
 systemctl daemon-reload 2>/dev/null
-rm /etc/systemd/system/pve-cert-copy.service
-rm /etc/systemd/system/pve-cert-copy.path
+rm /usr/local/bin/pve-cert-copy.sh
 rm /usr/local/bin/uninstall-pve-cert-copy.sh
 EOF
 
-chmod +x /usr/local/bin/uninstall-pve-cert-copy.sh
+chmod u+x /usr/local/bin/uninstall-pve-cert-copy.sh
+chmod u+x /usr/local/bin/pve-cert-copy.sh
 
 systemctl daemon-reload
-systemctl enable --now pve-cert-copy.path
 # Run once after installation
-systemctl start pve-cert-copy.service
+/usr/local/bin/pve-cert-copy.sh
